# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.CreateEventRequest:
    properties:
      category:
        enum:
        - seminar
        - workshop
        - lomba
        - konser
        type: string
      description:
        type: string
      end_date:
        type: string
      event_type:
        enum:
        - online
        - offline
        type: string
      is_uii_only:
        type: boolean
      location:
        type: string
      max_participants:
        minimum: 1
        type: integer
      registration_deadline:
        type: string
      start_date:
        type: string
      status:
        enum:
        - draft
        - published
        type: string
      title:
        type: string
      zoom_link:
        type: string
    required:
    - category
    - description
    - end_date
    - event_type
    - max_participants
    - registration_deadline
    - start_date
    - status
    - title
    type: object
...
  request.UpdateEventRequest:
    properties:
      category:
        enum:
        - seminar
        - workshop
        - lomba
        - konser
        type: string
      description:
        type: string
      end_date:
        type: string
      event_type:
        enum:
        - online
        - offline
        type: string
      is_uii_only:
        type: boolean
      location:
        type: string
      max_participants:
        minimum: 1
        type: integer
      registration_deadline:
        type: string
      start_date:
        type: string
      status:
        enum:
        - draft
        - published
        - ongoing
        - completed
        - cancelled
        type: string
      title:
        type: string
      zoom_link:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@eventcampus.com
    name: API Support
  description: |-
    REST API for Event Campus - Campus Event Management System

    Features:
    - User authentication (Register, Login)
    - Event management (Create, Read, Update, Delete)
    - Event registration
    - Attendance tracking
    - Whitelist/Organizer approval system
    - Email notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Event Campus API
  version: "1.0"
paths:
...
  /events:
    get:
      consumes:
      - application/json
      description: Get list of events with optional filters (category, status, event_type,
        search)
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by status (draft/published/ongoing/completed/cancelled)
        in: query
        name: status
        type: string
      - description: Filter by event type (online/offline/hybrid)
        in: query
        name: event_type
        type: string
      - description: Search by name or description
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Events retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get events
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all events
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Create a new event (organizer only). Poster can be uploaded separately.
      parameters:
      - description: Event details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Event created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or failed to create
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new event
      tags:
      - Events
  /events/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an event (organizer only)
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid event ID or deletion failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - Events
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific event
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid event ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Event not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get event by ID
      tags:
      - Events
    put:
      consumes:
      - application/json
      description: Update event details (organizer only)
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated event details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Event updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or update failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update event
      tags:
      - Events
...
  /events/{id}/poster:
    post:
      consumes:
      - multipart/form-data
      description: Upload poster image for an event (organizer only)
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Poster image (JPG/PNG)
        in: formData
        name: poster
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Poster uploaded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid file or upload failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload event poster
      tags:
      - Events
  /events/{id}/publish:
    post:
      consumes:
      - application/json
      description: Publish a draft event to make it visible to users
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event published successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid event ID or publish failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Publish event
      tags:
      - Events
...
  /events/my-events:
    get:
      consumes:
      - application/json
      description: Get all events created by authenticated organizer
      produces:
      - application/json
      responses:
        "200":
          description: Events retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get events
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my events
      tags:
      - Events
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"